requirements.txt
fastapi~=0.110.3
pydantic~=2.7.4
openai~=1.0.0
chromadb~=0.4.24
PyYAML~=6.0.1
requests~=2.32.3
uvicorn~=0.25.0
chainlit~=1.0.506
agents~=1.4.0
config~=0.5.1
typing~=3.7.4.3
pip~=23.3
attrs~=23.2.0
DateTime~=5.5
build~=1.0.3
wheel~=0.43.0
cryptography~=42.0.8
Pillow~=10.0.1
tornado~=6.4.1
Jinja2~=3.1.4
uuid~=1.30
setuptools~=65.5.1

